% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corncyc.gene.map}
\alias{corncyc.gene.map}
\title{CornCyc 8 gene frameIDs mapped to B73 v4 gene models.}
\format{A data frame in long format.}
\usage{
data(corncyc.gene.map)
}
\description{
This data is a direct dump from CornCyc 8.0.1.  It represents all genes in the database that have a
B73 v4 id associated with them.
}
\keyword{datasets}
