% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corncyc.pathway.reaction.map}
\alias{corncyc.pathway.reaction.map}
\title{CornCyc 8 pathway frameIDs mapped to reaction frameIDs.}
\format{A data frame in long format.}
\usage{
data(corncyc.pathway.reaction.map)
}
\description{
This data is a direct dump from CornCyc 8.0.1.  It represents all pathways in the database that have a
reaction frames associated with them, excluding superpathways.
}
\keyword{datasets}
