% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corncyc.reaction.gene.map}
\alias{corncyc.reaction.gene.map}
\title{CornCyc 8 reaction frameIDs mapped to gene frameIDs.}
\format{A data frame in long format.}
\usage{
data(corncyc.reaction.gene.map)
}
\description{
This data is a direct dump from CornCyc 8.0.1.  It represents all reactions in the database that have a
gene frames associated with them.
}
\keyword{datasets}
