% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.genes.classical_to_v3.map}
\alias{maize.genes.classical_to_v3.map}
\title{Classical gene models mapped to B73 v3 gene model ids}
\format{A data frame columns [classical, v3_id].}
\usage{
data(maize.genes.classical_to_v3.map)
}
\description{
Provided by MaizeGDB, mapping is 1 to 1.
}
\keyword{datasets}
