% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.genes.uniprot_to_v4.map}
\alias{maize.genes.uniprot_to_v4.map}
\title{UniProt Accessions mapped to B73 v4 gene model ids}
\format{A data frame columns [UniProt_Acc, v4_id].}
\usage{
data(maize.genes.uniprot_to_v4.map)
}
\description{
Downloaded from UniProt.  Parses "Gramene" cross-reference field for the B73 ids.  Mapping is may include many-to-many
relationships, but has unnested the B73 ids where possible.
}
\keyword{datasets}
