% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.v4.37.chromosomal.gene.positions}
\alias{maize.v4.37.chromosomal.gene.positions}
\title{Map from genes to all available transcripts.}
\format{A data frame with 4 columns and 39179 rows.}
\usage{
data(maize.v4.37.chromosomal.gene.positions)
}
\description{
Parsed from the zea mays b73 v4.37 gff3 file available at ftp://ftp.ensemblgenomes.org/pub/release-37/plants/gff3/zea_mays/Zea_mays.AGPv4.37.gff3.gz.
Includes the 10 primary, Mt, and Pt chromosomes as well as unplaced scaffolds. Lists gene models and their start and end positions.
Contains some v3 ID's.
}
\keyword{datasets}
