% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.transcript.map}
\alias{gene.transcript.map}
\title{Map from genes to all available transcripts.}
\format{A data frame in long format.}
\usage{
data(gene.transcript.map)
}
\description{
Parsed from the zea mays b73 v4.37 gff3 file.  Excludes orphan exons. Excludes gene/transcripts with ID's beginning with MI, zma, or ENSRN, which
are all not standard v4 identifiers.
}
\keyword{datasets}
