% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.genes.v3_to_v4.map}
\alias{maize.genes.v3_to_v4.map}
\title{Unpublished map between B73 v3 and v4 gene model ids}
\format{A data frame columns [v3_id, v4_id].}
\usage{
data(maize.genes.v3_to_v4.map)
}
\description{
Map between B73 v3 and v4 gene model ids generated by Maggie using SynMap. This includes merged v3 models
but not split v3 models.  B73 V3 ids are of the form GRMZM and B73 v4 ids are of the form Zm.  V3 also
includes some "AC" ids.  These legacy ids were generated from BAC's.  The suffix on the AC ids, "FG###"
refers to the sequence the expressing segment appeared on that particular BAC.
}
\keyword{datasets}
